abstract job monitor
